requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
groq==0.8.0
httpx==0.27.0
langgraph==0.1.19
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.7.1
jinja2==3.1.4
aiofiles==23.2.1